# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "L5 Swagger API",
        "version": "1.0.0"
    },
    "host": "localhost:800/noithatkagu/public",
    "basePath": "/api",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Token will be return after login success",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Login",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "User registration",
                "description": "User registration",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User information",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                },
                                "password": {
                                    "type": "string"
                                },
                                "confirmPassword": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "phone": {
                                    "type": "string"
                                },
                                "address": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/change-password": {
            "put": {
                "tags": [
                    "Auth"
                ],
                "summary": "Change password",
                "description": "Change password",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Enter recent password and new password with confirm",
                        "required": true,
                        "schema": {
                            "properties": {
                                "currentPassword": {
                                    "type": "string"
                                },
                                "newPassword": {
                                    "type": "string"
                                },
                                "confirmNewPassword": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                },
                "security": [
                    {
                        "jwt": []
                    }
                ]
            }
        },
        "/auth/request/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Send request reset password",
                "description": "Send email attach token",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Email to reset",
                        "required": true,
                        "schema": {
                            "properties": {
                                "email": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/auth/accept/reset-password": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Accept request reset password",
                "description": "Finish reset password",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Reset Password",
                        "required": true,
                        "schema": {
                            "properties": {
                                "token": {
                                    "type": "string"
                                },
                                "newPassword": {
                                    "type": "string"
                                },
                                "confirmNewPassword": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "jwt": {
            "type": "apiKey",
            "description": "Example Format: Bearer Your-Token",
            "name": "authorization",
            "in": "header"
        }
    }
}